code is now divided between the main functionality, functions to get/parse data, and general utilities
